Implemented the NumpyPointCloud class throughout all files, testing tbd.
modules/np_pointcloud.py:
Added shortcuts for getting the point coordinate, normals and sigma 
fields.

If only one column is returned in self.get_fields(), it is reshaped to 1 
dimension only, so indexing can be done.
Added check for number of labels

All files:
Updated most of the functions to work with NumpyPointClouds, exept those 
that don't return clouds and do not change the fields of the clouds 
given.

Removed old code (functions, comments), restructured this and that, 
sparingly updated docstrings